changed int to long

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[]args){
-        int[] sales = new int[] {4,2,100,400};
+        long[] sales = new long[] {4,2,100,400};
         SalesManager salesManager = new SalesManager(sales);
-        int max = salesManager.max();
+        long max = salesManager.max();
         System.out.println(max);
 
     }
